fix(fill): sparse_anchoring honeycomb density via f64 product form
The bridge-over-infill anchoring pass constructed the 3D honeycomb
density with the pure-f32 product (0.01 * f32 density), one form left
behind by the earlier Fill.cpp:1283 port. Every sibling pattern in the
same match arm uses (0.01_f64 * f64::from(..)) as f32. The 1-ulp-low
density shifted that invocation's grid to 3579535.68 (correct:
3579535.104) and zScale to 1.789767840 (correct: 1.789767552).

Verified against GT %.17g dumps: all 39 invocations now compute
grid=3579535.10424945643171668 and the per-z triWave chain matches
Orca bit-identically (max abs diff 0 across 39 invocations).

+++ crates/ares-core/src/project_slice/prepare_infill/bridge_over_infill/sparse_anchoring.rs
@@ -185,7 +185,7 @@ pub(in crate::project_slice) fn generate_sparse_infill_polylines_for_anchoring(
                     spacing: fill.params.spacing,
                     overlap: 0.0,
                     angle: fill.params.angle,
-                    density: 0.01 * fill.params.density,
+                    density: (0.01_f64 * f64::from(fill.params.density)) as f32,
                     multiline: fill.params.multiline,
                     anchor_length: fill.params.anchor_length,
                     anchor_length_max: fill.params.anchor_length_max,
